Windows build fixed

+++ toolkit/Source/src/engine/Entity/ETHEntityArray.cpp
@@ -69,18 +69,18 @@ const ETHEntityRawPtr &ETHEntityArray::operator[] (const unsigned int index) con
 
 ETHEntityRawPtr &ETHEntityArray::operator[] (const unsigned long index)
 {
-	return (*this)[index];
+	return (*this)[static_cast<unsigned int>(index)];
 }
 
 const ETHEntityRawPtr &ETHEntityArray::operator[] (const unsigned long index) const
 {
-	return (*this)[index];
+	return (*this)[static_cast<unsigned int>(index)];
 }
 
 ETHEntityArray &ETHEntityArray::operator+=(const ETHEntityArray &other)
 {
 	const unsigned int nSize = other.size();
-	for (unsigned int t=0; t<nSize; t++)
+	for (unsigned int t = 0; t < nSize; t++)
 	{
 		push_back(other[t]);
 	}
@@ -92,7 +92,7 @@ void ETHEntityArray::clear()
 	const unsigned int nSize = size();
 	if (nSize)
 	{
-		for (unsigned int t=0; t<nSize; t++)
+		for (unsigned int t = 0; t < nSize; t++)
 		{
 			if (m_vector[t])
 			{
@@ -107,7 +107,7 @@ ETHEntityArray &ETHEntityArray::operator=(const ETHEntityArray &other)
 {
 	clear();
 	m_vector.resize(other.size());
-	for (unsigned int t=0; t<other.size(); t++)
+	for (unsigned int t = 0; t < other.size(); t++)
 	{
 		m_vector[t] = other[t];
 	}
